relocated staffSig component to BySupervisor dir and fixed imports across app
